+++ tests/Feature/Management/Fakes/ManagedModelFakeSecond.php
@@ -6,14 +6,14 @@ use Illuminate\Support\Facades\Schema;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Schema\Blueprint;
 use Thinktomorrow\Chief\States\PageState;
-use Spatie\MediaLibrary\HasMedia\HasMedia;
 use Thinktomorrow\Chief\Management\ManagedModel;
-use Thinktomorrow\AssetLibrary\Traits\AssetTrait;
 use Thinktomorrow\Chief\States\Publishable\Publishable;
+use Thinktomorrow\AssetLibrary\AssetTrait;
+use Thinktomorrow\AssetLibrary\HasAsset;
 use Thinktomorrow\Chief\Concerns\Translatable\Translatable;
 use Thinktomorrow\Chief\Concerns\Translatable\TranslatableContract;
 
-class ManagedModelFakeSecond extends Model implements  ManagedModel, TranslatableContract, HasMedia
+class ManagedModelFakeSecond extends Model implements  ManagedModel, TranslatableContract, HasAsset
 {
     use Translatable, \Astrotomic\Translatable\Translatable, AssetTrait, Publishable;