+++ tests/Feature/Management/Fakes/ManagerFake.php
@@ -13,8 +13,6 @@ use Thinktomorrow\Chief\Management\Manager;
 
 class ManagerFake extends AbstractManager implements Manager
 {
-
-
     public function fields(): Fields
     {
         return parent::fields()->add(