+++ tests/Feature/Media/UploadMediaTest.php
@@ -4,6 +4,7 @@ namespace Thinktomorrow\Chief\Tests\Feature\Pages\Media;
 
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Route;
+use Thinktomorrow\AssetLibrary\Asset;
 use Thinktomorrow\Chief\Pages\Single;
 use Thinktomorrow\Chief\Modules\Module;
 use Thinktomorrow\Chief\Tests\TestCase;
@@ -11,7 +12,9 @@ use Thinktomorrow\Chief\Media\MediaType;
 use Thinktomorrow\Chief\Pages\PageManager;
 use Thinktomorrow\Chief\Management\Register;
 use Thinktomorrow\Chief\Tests\Fakes\MediaModule;
+use Thinktomorrow\AssetLibrary\Application\AddAsset;
 use Thinktomorrow\Chief\Tests\Fakes\UploadMediaManager;
+use Thinktomorrow\AssetLibrary\Application\AssetUploader;
 use Thinktomorrow\Chief\Tests\Feature\Pages\PageFormParams;
 use Thinktomorrow\Chief\Tests\Fakes\UploadMediaModuleManager;
 
@@ -52,8 +55,7 @@ class UploadMediaTest extends TestCase
                 ]
             ]));
 
-        $this->assertTrue($page->fresh()->hasFile(MediaType::HERO));
-        $this->assertCount(1, $page->fresh()->getAllFiles(MediaType::HERO));
+        $this->assertCount(1, $page->fresh()->assets(MediaType::HERO));
     }
 
     /** @test */
@@ -75,8 +77,7 @@ class UploadMediaTest extends TestCase
                 ]
             ]);
 
-        $this->assertTrue($module->fresh()->hasFile(MediaType::HERO));
-        $this->assertCount(1, $module->fresh()->getAllFiles(MediaType::HERO));
+        $this->assertCount(1, $module->fresh()->assets(MediaType::HERO));
     }
 
     /** @test */
@@ -100,17 +101,16 @@ class UploadMediaTest extends TestCase
                 ]
             ]));
 
-        $this->assertTrue($page->fresh()->hasFile(MediaType::DOCUMENT));
-        $this->assertCount(1, $page->getAllFiles(MediaType::DOCUMENT));
+        $this->assertCount(1, $page->assets(MediaType::DOCUMENT));
     }
 
     /** @test */
     public function an_asset_can_be_replaced()
     {
         $page = Single::create();
-        $page->addFile(UploadedFile::fake()->image('image.png'), MediaType::HERO);
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), MediaType::HERO, 'nl');
 
-        $existing_asset = $page->getAllFiles(MediaType::HERO)->first();
+        $existing_asset = $page->assets(MediaType::HERO)->first();
 
         // Replace asset
         $this->asAdmin()
@@ -125,19 +125,20 @@ class UploadMediaTest extends TestCase
             ]));
 
         // Assert replacement took place
-        $this->assertCount(1, $page->fresh()->getAllFiles(MediaType::HERO));
-        $this->assertStringContainsString('tt-favicon.png', $page->fresh()->getFileUrl(MediaType::HERO));
+        $this->assertCount(1, $page->fresh()->assets(MediaType::HERO));
+        $this->assertCount(2, Asset::all());
+
+        $this->assertStringContainsString('tt-favicon.png', $page->fresh()->asset(MediaType::HERO, 'nl')->filename());
     }
 
     /** @test */
     public function an_asset_can_be_removed()
     {
         $page = Single::create();
-        $page->addFile(UploadedFile::fake()->image('image.png'), MediaType::HERO);
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), MediaType::HERO, 'nl');
 
         // Assert Image is there
-        $this->assertTrue($page->hasFile(MediaType::HERO));
-        $this->assertCount(1, $page->getAllFiles(MediaType::HERO));
+        $this->assertCount(1, $page->assets(MediaType::HERO));
 
         // Remove asset
         $response = $this->asAdmin()
@@ -145,26 +146,25 @@ class UploadMediaTest extends TestCase
                 'files' => [
                     MediaType::HERO => [
                         'delete' => [
-                            $page->getAllFiles(MediaType::HERO)->first()->id,
+                            $page->assets(MediaType::HERO)->first()->id,
                         ]
                     ]
                 ]
             ]));
 
         // Assert Image is no longer there
-        $this->assertFalse($page->fresh()->hasFile(MediaType::HERO));
-        $this->assertCount(0, $page->fresh()->getAllFiles());
+        $this->assertCount(0, $page->fresh()->assets());
+        $this->assertCount(1, Asset::all());
     }
 
     /** @test */
     public function an_asset_can_be_removed_and_uploaded()
     {
         $page = Single::create();
-        $page->addFile(UploadedFile::fake()->image('image.png'), MediaType::HERO);
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), MediaType::HERO, 'nl');
 
         // Assert Image is there
-        $this->assertTrue($page->hasFile(MediaType::HERO));
-        $this->assertCount(1, $page->getAllFiles(MediaType::HERO));
+        $this->assertCount(1, $page->assets(MediaType::HERO));
 
         // Remove asset
         $response = $this->asAdmin()
@@ -172,7 +172,7 @@ class UploadMediaTest extends TestCase
                 'files' => [
                     MediaType::HERO => [
                         'delete' => [
-                            $page->getAllFiles(MediaType::HERO)->first()->id,
+                            $page->assets(MediaType::HERO)->first()->id,
                         ],
                         'new' => [
                             UploadedFile::fake()->image('image.png')
@@ -181,7 +181,7 @@ class UploadMediaTest extends TestCase
                 ],
             ]);
 
-        $this->assertCount(1, $page->fresh()->getAllFiles());
+        $this->assertCount(1, $page->fresh()->assets());
     }
 
     /** @test */
@@ -204,7 +204,7 @@ class UploadMediaTest extends TestCase
                 ]
             ]));
 
-        $this->assertEquals('tt-favicon.png', $page->getFilename(MediaType::HERO));
+        $this->assertEquals('tt-favicon.png', $page->asset(MediaType::HERO)->filename());
     }
 
     /** @test */
@@ -231,8 +231,8 @@ class UploadMediaTest extends TestCase
                 ]
             ]));
 
-        $this->assertEquals('tt-favicon-nl.png', $page->getFilename('seo_image', 'nl'));
-        $this->assertEquals('tt-favicon-en.png', $page->getFilename('seo_image', 'en'));
+        $this->assertEquals('tt-favicon-nl.png', $page->asset('seo_image', 'nl')->filename());
+        $this->assertEquals('tt-favicon-en.png', $page->asset('seo_image', 'en')->filename());
     }
 
     /** @test */
@@ -240,9 +240,9 @@ class UploadMediaTest extends TestCase
     {
         app(Register::class)->register(PageManager::class, Single::class);
         $page = Single::create();
-        $page->addFile(UploadedFile::fake()->image('image.png'), 'seo_image', 'nl');
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), 'seo_image', 'nl');
 
-        $existing_asset_nl = $page->getAllFiles('seo_image', 'nl')->first();
+        $existing_asset_nl = $page->assets('seo_image', 'nl')->first();
 
         $this->asAdmin()
             ->put(route('chief.back.managers.update', ['singles', $page->id]), $this->validUpdatePageParams([
@@ -262,8 +262,8 @@ class UploadMediaTest extends TestCase
                 ]
             ]));
 
-        $this->assertEquals('tt-favicon-nl.png', $page->fresh()->getFilename('seo_image', 'nl'));
-        $this->assertEquals('tt-favicon-en.png', $page->fresh()->getFilename('seo_image', 'en'));
+        $this->assertEquals('tt-favicon-nl.png', $page->fresh()->asset('seo_image', 'nl')->filename());
+        $this->assertEquals('tt-favicon-en.png', $page->fresh()->asset('seo_image', 'en')->filename());
     }
 
     /** @test */
@@ -271,9 +271,9 @@ class UploadMediaTest extends TestCase
     {
         app(Register::class)->register(PageManager::class, Single::class);
         $page = Single::create();
-        $page->addFile(UploadedFile::fake()->image('image.png'), 'seo_image', 'en');
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), 'seo_image', 'en');
 
-        $existing_asset_en = $page->getAllFiles('seo_image', 'en')->first();
+        $existing_asset_en = $page->assets('seo_image', 'en')->first();
 
         $this->asAdmin()
             ->put(route('chief.back.managers.update', ['singles', $page->id]), $this->validUpdatePageParams([
@@ -292,9 +292,9 @@ class UploadMediaTest extends TestCase
                     ]
                 ]
             ]));
-
-        $this->assertEquals('tt-favicon-nl.png', $page->fresh()->getFilename('seo_image', 'nl'));
-        $this->assertEquals('tt-favicon-nl.png', $page->fresh()->getFilename('seo_image', 'en'));
+        
+        $this->assertEquals('tt-favicon-nl.png', $page->refresh()->asset('seo_image', 'nl')->filename());
+        $this->assertEquals('tt-favicon-nl.png', $page->asset('seo_image', 'en')->filename());
     }
 
     /** @test */
@@ -302,10 +302,10 @@ class UploadMediaTest extends TestCase
     {
         $this->disableExceptionHandling();
         $page = Single::create();
-        $page->addFile(UploadedFile::fake()->image('image.png'), MediaType::HERO);
-        $page->addFile(UploadedFile::fake()->image('image2.png'), MediaType::HERO);
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), MediaType::HERO, 'nl');
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image2.png'), MediaType::HERO, 'nl');
 
-        $images = $page->fresh()->getAllFiles(MediaType::HERO);
+        $images = $page->fresh()->assets(MediaType::HERO);
 
         $this->asAdmin()
             ->put(route('chief.back.managers.update', ['singles', $page->id]), $this->validUpdatePageParams([
@@ -316,7 +316,7 @@ class UploadMediaTest extends TestCase
                 ]
             ]));
 
-        $assetIds = $page->fresh()->getAllFiles(MediaType::HERO)->pluck('id')->toArray();
+        $assetIds = $page->fresh()->assets(MediaType::HERO)->pluck('id')->toArray();
 
         $this->assertEquals([$images->last()->id, $images->first()->id], $assetIds);
     }
@@ -326,30 +326,141 @@ class UploadMediaTest extends TestCase
     {
         $this->disableExceptionHandling();
         $page = Single::create();
-        $page->addFile(UploadedFile::fake()->image('image.png'), MediaType::HERO, 'nl');
-        $page->addFile(UploadedFile::fake()->image('image2.png'), MediaType::HERO, 'nl');
-        $page->addFile(UploadedFile::fake()->image('image3.png'), MediaType::HERO, 'en');
-        $page->addFile(UploadedFile::fake()->image('image4.png'), MediaType::HERO, 'en');
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), MediaType::HERO, 'nl');
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image2.png'), MediaType::HERO, 'nl');
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image3.png'), MediaType::HERO, 'en');
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image4.png'), MediaType::HERO, 'en');
 
-        $images = $page->assets()->get();
+        $nl_images = $page->assets(MediaType::HERO, 'nl');
+        $en_images = $page->assets(MediaType::HERO, 'en');
+
         $this->asAdmin()
             ->put(route('chief.back.managers.update', ['singles', $page->id]), $this->validUpdatePageParams([
                 'filesOrder' => 
                 [
                     'nl' => [
-                        'files-' . MediaType::HERO => $images[1]->id . ',' . $images[0]->id,
+                        'files-' . MediaType::HERO => $nl_images[1]->id . ',' . $nl_images[0]->id,
                     ],
                     'en' => [
-                        'files-' . MediaType::HERO => $images[3]->id . ',' . $images[2]->id
+                        'files-' . MediaType::HERO => $en_images[3]->id . ',' . $en_images[2]->id
                     ]
                 ]
             ]));
 
-        $newImagesSorted = $page->fresh()->assets()->get();
+        $nl_newImagesSorted = $page->refresh()->assets(MediaType::HERO, 'nl');
+        $en_newImagesSorted = $page->assets(MediaType::HERO, 'en');
 
-        $this->assertEquals($images[1]->id, $newImagesSorted[0]->id);
-        $this->assertEquals($images[0]->id, $newImagesSorted[1]->id);
-        $this->assertEquals($images[3]->id, $newImagesSorted[2]->id);
-        $this->assertEquals($images[2]->id, $newImagesSorted[3]->id);
+        $this->assertEquals($nl_images[1]->id, $nl_newImagesSorted[0]->id);
+        $this->assertEquals($nl_images[0]->id, $nl_newImagesSorted[1]->id);
+        $this->assertEquals($en_images[3]->id, $en_newImagesSorted[2]->id);
+        $this->assertEquals($en_images[2]->id, $en_newImagesSorted[3]->id);
+    }
+
+    /** @test */
+    public function an_existing_asset_can_be_added()
+    {
+        $page = Single::create();
+        $existing_asset = AssetUploader::upload(UploadedFile::fake()->image('image.png'));
+
+        $this->asAdmin()
+            ->put(route('chief.back.managers.update', ['singles', $page->id]), $this->validUpdatePageParams([
+                'files' => [
+                    MediaType::HERO => [
+                        'new' => [
+                            $existing_asset->id,
+                        ]
+                    ]
+                ]
+            ]));
+
+        // Assert replacement took place
+        $this->assertCount(1, $page->fresh()->assets(MediaType::HERO));
+
+        $this->assertEquals($existing_asset->url(), $page->fresh()->asset(MediaType::HERO, 'nl')->url());
+    }
+    
+    /** @test */
+    public function an_existing_asset_can_be_added_as_translation()
+    {
+        $page = Single::create();
+        $existing_asset = AssetUploader::upload(UploadedFile::fake()->image('image.png'));
+
+        $this->asAdmin()
+            ->put(route('chief.back.managers.update', ['singles', $page->id]), $this->validUpdatePageParams([
+                'files' => [
+                    MediaType::HERO => [
+                        'en' => [
+                            'new' => [
+                                $existing_asset->id
+                            ]
+                        ]
+                    ]
+                ]
+            ]));
+
+        $this->assertCount(1, $page->fresh()->assets(MediaType::HERO, 'en'));
+
+        $this->assertEquals($existing_asset->url(), $page->fresh()->asset(MediaType::HERO, 'en')->url());
+    }
+
+    /** @test */
+    public function it_can_link_an_asset_to_multiple_pages()
+    {
+        $page = Single::create();
+        $page2 = Single::create();
+        $existing_asset = AssetUploader::upload(UploadedFile::fake()->image('image.png'));
+
+        $this->asAdmin()
+            ->put(route('chief.back.managers.update', ['singles', $page->id]), $this->validUpdatePageParams([
+                'files' => [
+                    MediaType::HERO => [
+                        'en' => [
+                            'new' => [
+                                $existing_asset->id
+                            ]
+                        ]
+                    ]
+                ]
+            ]));
+
+        $this->asAdmin()
+            ->put(route('chief.back.managers.update', ['singles', $page2->id]), $this->validUpdatePageParams([
+                'trans' => [
+                    'nl' => [
+                        'slug' => 'page 2'
+                    ]
+                ],
+                'url-slugs' => [
+                    'nl' => 'slug',
+                    'en' => 'slug',
+                ],
+                'files' => [
+                    MediaType::HERO => [
+                        'en' => [
+                            'new' => [
+                                $existing_asset->id
+                            ]
+                        ]
+                    ]
+                ]
+            ]));
+
+        $this->assertCount(1, $page->refresh()->assets(MediaType::HERO, 'en'));
+        $this->assertCount(1, $page2->refresh()->assets(MediaType::HERO, 'en'));
+
+        $this->assertEquals($existing_asset->url(), $page->asset(MediaType::HERO, 'en')->url());
+        $this->assertEquals($existing_asset->url(), $page2->asset(MediaType::HERO, 'en')->url());
+    }
+
+    /** @test */
+    public function removing_a_model_with_asset_unlinks_the_asset()
+    {
+        $page = Single::create();
+        app(AddAsset::class)->add($page, UploadedFile::fake()->image('image.png'), MediaType::HERO, 'nl');
+
+        $this->asAdmin()
+            ->delete('/admin/manage/singles/'.$page->id);
+
+        $this->assertCount(1, Asset::all());
     }
 }